% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDCweek.R
\name{CDCweek}
\alias{CDCweek}
\title{CDC week and year}
\usage{
CDCweek(date)
}
\arguments{
\item{date}{vector which can be coerced to class \code{Date}}
}
\value{
data.frame with elements day (of the week), CDC week, and CDC year
}
\description{
This function returns the CDC week and year for the date(s) provided.
}
\examples{
y = as.Date(paste(1999:2011, "-12-31", sep = ""))
cbind(y, CDCweek(y))
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\references{
\url{http://stackoverflow.com/questions/28238438/epidemiological-curve-using-r-and-epitools-gives-wrong-epiweeks}
}
\seealso{
\code{\link{get_start_date}}, \code{\link{CDCweekday}}
}

