% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDCweek2Date.R
\name{CDCweek2Date}
\alias{CDCweek2Date}
\title{Convert CDCweek to Date}
\usage{
CDCweek2Date(CDCyear, CDCweek, CDCday = NULL)
}
\arguments{
\item{CDCyear}{numeric vector of years}

\item{CDCweek}{numeric vector of weeks}

\item{CDCday}{numeric vector of days, defaults to a vector of 1s}
}
\value{
Date vector of actual dates
}
\description{
Computes the Date from the CDC year, week, and day.
}
\examples{
CDCweek2Date(2015,36,3)
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\seealso{
\code{\link{CDCweek}}
}

