% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDCweekday.R
\name{CDCweekday}
\alias{CDCweekday}
\title{Day of week according to CDC}
\usage{
CDCweekday(date)
}
\arguments{
\item{date}{vector which can be coerced to class \code{Date}}
}
\value{
vector of weekdays as a factor
}
\description{
This function returns the weekday of a given date according to MMWR.
}
\examples{
y <- as.Date(paste(1999:2011, "-12-31", sep = ""))
data.frame(date = format(y), CDCweekday = CDCweekday(y))
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\seealso{
\code{\link{CDCweek}}
}

