% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMWRweek2Date.R
\name{MMWRweek2Date}
\alias{MMWRweek2Date}
\title{Convert MMWRweek to Date}
\usage{
MMWRweek2Date(MMWRyear, MMWRweek, MMWRday = NULL)
}
\arguments{
\item{MMWRyear}{numeric vector of years}

\item{MMWRweek}{numeric vector of weeks}

\item{MMWRday}{numeric vector of days, defaults to a vector of 1s}
}
\value{
Date vector of dates associated with MMWR year, week, and day
}
\description{
Computes the Date from the MMWR year, week, and day.
}
\examples{
MMWRweek2Date(MMWRyear=2015,MMWRweek=36,MMWRday=3)
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\seealso{
\code{\link{MMWRweek}}
}

