% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_date.R
\name{start_date}
\alias{start_date}
\title{Find start date for a calendar year}
\usage{
start_date(year)
}
\arguments{
\item{year}{integer vector of four digit years}
}
\value{
Date vector for January 1st of the calendar year
}
\description{
Finds the state date given a numeric calendar year
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\seealso{
\code{\link{get_start_date}}
}

