% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_start_date.R
\name{get_start_date}
\alias{get_start_date}
\title{Finds the start date for the year associated with date.}
\usage{
get_start_date(date)
}
\arguments{
\item{date}{Date vector}
}
\value{
Date vector for start date of CDC year associated with date
}
\description{
Values for MMWR week range from 1 to 53, although most years consist of
52 weeks. The first day of any MMWR week is Sunday. MMWR week numbering
is sequential beginning with 1 and incrementing with each week to a maximum
of 52 or 53. MMWR week #1 of an MMWR year is the first week of the year
that has at least four days in the calendar year. For example, if January
1 occurs on a Sunday, Monday, Tuesday or Wednesday, the calendar week that
includes January 1 would be MMWR week #1. If January 1 occurs on a Thursday,
Friday, or Saturday, the calendar week that includes January 1 would be the
last MMWR week of the previous year (#52 or #53). Because of this rule,
December 29, 30, and 31 could potentially fall into MMWR week #1 of the
following MMWR year. This function returns the start date for the year
associated with this date.
}
\examples{
get_start_date(as.Date('2015-01-01')+ -10:10)
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
\references{
\url{http://stackoverflow.com/questions/28238438/epidemiological-curve-using-r-and-epitools-gives-wrong-epiweeks}
}
\seealso{
\code{\link{start_date}}
}

