% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_date.R
\name{get_start_date}
\alias{get_start_date}
\title{Finds the start date for the year associated with date.}
\usage{
get_start_date(date)
}
\arguments{
\item{date}{vector which can be coerced to class \code{Date}}
}
\value{
Date vector for start date of MMWR year associated with date
}
\description{
Finds the start date for the year associated with date.
}
\seealso{
\code{\link{start_date}}
}
\author{
Jarad Niemi \email{niemi@iastate.edu}
}
